#56 💷  refactor : components 폴더 내 myAsset 폴더 생성
